Pleased Giraffe Fix 2.0
I uploaded a complete bollox shortly after the Pleased Giraffe update -
the dev version was working and I managed to screw up the public build
two month ago. Now it is fixed.
As well as I changed the key for the flight-ability, it is now on
H(Special Key 3) and not anymore on F, which should prevent any key
conflicts.
Cheers.